feat: guess API 추가

+++ src/main/java/com/keeper/homepage/domain/game/dto/req/BaseballStartRequest.kt
@@ -8,5 +8,5 @@ const val MAX_TITLE_LENGTH = 10000L
 data class BaseballStartRequest(
     @field:Positive
     @field:Max(MAX_TITLE_LENGTH)
-    val bettingPoint: Long,
+    val bettingPoint: Int,
 )
